payment mthod

+++ database/migrations/2026_02_14_003636_reservations.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('reservations')) {
+            return;
+        }
+
         Schema::create('reservations', function(Blueprint $table){
             $table->id();
             $table->foreignId('client_id')->constrained()->onDelete('restrict');
